kisan ai/requirements.txt
django==5.1.2
psycopg2==2.9.10
langchain==0.3.3
google-generativeai==0.8.3
beautifulsoup4==4.12.3
requests==2.32.3
torch==2.4.1
joblib==1.4.2
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.0